requirements.txt
Flask==2.2.5
gunicorn==20.1.0
werkzeug==2.2.3
pyspellchecker==0.7.2
language_tool_python==2.7.1
groq
python-dotenv
gunicorn